fix(text): improve error handling in napi async work and loop variable type
- Add error checking for snprintf_s in napi_async_work.h to prevent potential buffer issues
- Change loop variable type from int to size_t in font_parser.cpp for better consistency and safety

+++ frameworks/text/service/texgine/src/font_parser.cpp
@@ -211,7 +211,7 @@ void FontParser::ProcessTable(const NameTable* nameTable, FontParser::FontDescri
     auto count = nameTable->count.Get();
     auto storageOffset = nameTable->storageOffset.Get();
     const char* stringStorage = reinterpret_cast<const char*>(nameTable) + storageOffset;
-    for (int i = 0; i < count && (sizeof(NameTable) + (i + 1) * sizeof(NameRecord)) <= size; ++i) {
+    for (size_t i = 0; i < count && (sizeof(NameTable) + (i + 1) * sizeof(NameRecord)) <= size; ++i) {
         if (nameTable->nameRecord[i].stringOffset.Get() == 0 && nameTable->nameRecord[i].length.Get() == 0) {
             continue;
         }
